Restructured project file structures and added more classes for future work
