--- src/main/java/org/example/cyberwatch/controller/TicketController.java
@@ -1,4 +1,0 @@
-package org.example.cyberwatch.controller;
-
-public class TicketController {
-}
