+++ src/main/java/org/example/cyberwatch/features/form/controller/FormController.java
@@ -1,0 +1,7 @@
+package org.example.cyberwatch.features.form.controller;
+
+import org.springframework.stereotype.Controller;
+
+@Controller
+public class FormController {
+}
