+++ src/main/java/org/example/cyberwatch/features/form/model/EmploymentForm.java
@@ -1,0 +1,7 @@
+package org.example.cyberwatch.features.form.model;
+
+//Could it be an idea that the EmployeeForm needs to be approved by an Manager, like a signature on a paper form?
+// Then we could have a status field in the EmployeeForm with the following states:
+//DRAFT -> SUBMITTED -> APPROVED -> COMPLETED -> REJECTED
+public class EmploymentForm {
+}
