+++ src/main/java/org/example/cyberwatch/features/form/model/FormDTO.java
@@ -1,0 +1,4 @@
+package org.example.cyberwatch.features.form.model;
+
+public class FormDTO {
+}
