+++ src/main/java/org/example/cyberwatch/features/form/repository/FormRepository.java
@@ -1,0 +1,9 @@
+package org.example.cyberwatch.features.form.repository;
+
+import org.example.cyberwatch.features.form.model.ReportForm;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface FormRepository extends JpaRepository<ReportForm, Long> {
+}
