+++ src/main/java/org/example/cyberwatch/features/form/service/FormService.java
@@ -1,0 +1,4 @@
+package org.example.cyberwatch.features.form.service;
+
+public class FormService {
+}
