+++ src/main/java/org/example/cyberwatch/features/staff/controller/StaffController.java
@@ -1,0 +1,7 @@
+package org.example.cyberwatch.features.staff.controller;
+
+import org.springframework.stereotype.Controller;
+
+@Controller
+public class StaffController {
+}
