+++ src/main/java/org/example/cyberwatch/features/staff/exception/StaffNotFoundException.java
@@ -1,0 +1,4 @@
+package org.example.cyberwatch.features.staff.exception;
+
+public class StaffNotFoundException {
+}
