+++ src/main/java/org/example/cyberwatch/features/staff/model/Management.java
@@ -1,0 +1,4 @@
+package org.example.cyberwatch.features.staff.model;
+
+public class Management {
+}
