+++ src/main/java/org/example/cyberwatch/features/staff/model/StaffDTO.java
@@ -1,0 +1,4 @@
+package org.example.cyberwatch.features.staff.model;
+
+public class StaffDTO {
+}
