+++ src/main/java/org/example/cyberwatch/features/staff/service/StaffService.java
@@ -1,0 +1,4 @@
+package org.example.cyberwatch.features.staff.service;
+
+public class StaffService {
+}
