+++ src/main/java/org/example/cyberwatch/features/ticket/controller/TicketController.java
@@ -1,0 +1,4 @@
+package org.example.cyberwatch.features.ticket.controller;
+
+public class TicketController {
+}
