+++ src/main/java/org/example/cyberwatch/features/ticket/exception/TicketNotFoundException.java
@@ -1,0 +1,4 @@
+package org.example.cyberwatch.features.ticket.exception;
+
+public class TicketNotFoundException {
+}
