+++ src/main/java/org/example/cyberwatch/features/ticket/model/Ticket.java
@@ -1,0 +1,4 @@
+package org.example.cyberwatch.features.ticket.model;
+
+public class Ticket {
+}
