+++ src/main/java/org/example/cyberwatch/features/ticket/model/TicketDTO.java
@@ -1,0 +1,4 @@
+package org.example.cyberwatch.features.ticket.model;
+
+public class TicketDTO {
+}
