+++ src/main/java/org/example/cyberwatch/features/ticket/model/entitys/staff/Staff.java
@@ -1,0 +1,106 @@
+package org.example.cyberwatch.features.ticket.model.entitys.staff;
+
+import jakarta.persistence.*;
+import jakarta.validation.constraints.Email;
+import org.example.cyberwatch.shared.model.enums.Department;
+
+// Represents an employee in the system, base entity for all staff with personal information. Linked 1:1 to HR/Management/Consultant roles.
+//Differs from Employee form wich is the process/form HR uses to create a new employee.
+@Entity
+public class Staff {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "employee_id")
+    Long id;
+
+    @Column(name = "social_security_number", nullable = false, unique = true)
+    String socialSecurityNumber;
+
+    @Column(name = "first_name")
+    String firstName;
+
+    @Column(name = "last_name")
+    String lastName;
+
+    @Email(message = "Email should be valid")
+    String email;
+
+    @Column(name = "phone_number")
+    String phoneNumber;
+
+    String role;
+
+    @Enumerated(EnumType.STRING)
+    Department department;
+
+
+    public Staff() {
+
+    }
+
+    public String getSocialSecurityNumber() {
+        return socialSecurityNumber;
+    }
+
+    public void setSocialSecurityNumber(String socialSecurityNumber) {
+        this.socialSecurityNumber = socialSecurityNumber;
+    }
+
+    public String getRole() {
+        return role;
+    }
+
+    public void setRole(String role) {
+        this.role = role;
+    }
+
+    public Department getDepartment() {
+        return department;
+    }
+
+    public void setDepartment(Department department) {
+        this.department = department;
+    }
+
+    public String getFirstName() {
+        return firstName;
+    }
+
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getPhoneNumber() {
+        return phoneNumber;
+    }
+
+    public void setPhoneNumber(String phoneNumber) {
+        this.phoneNumber = phoneNumber;
+    }
+
+    public String getLastName() {
+        return lastName;
+    }
+
+    public void setLastName(String lastName) {
+        this.lastName = lastName;
+    }
+
+
+}
