+++ src/main/java/org/example/cyberwatch/features/ticket/repository/TicketRepository.java
@@ -1,0 +1,8 @@
+package org.example.cyberwatch.features.ticket.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface TicketRepository extends JpaRepository<Ticket, Long> {
+}
