+++ src/main/java/org/example/cyberwatch/features/ticket/service/TicketService.java
@@ -1,0 +1,4 @@
+package org.example.cyberwatch.features.ticket.service;
+
+public class TicketService {
+}
