--- src/main/java/org/example/cyberwatch/model/entitys/forms/EmploymentForm.java
@@ -1,7 +1,0 @@
-package org.example.cyberwatch.model.entitys.forms;
-
-//Could it be an idea that the EmployeeForm needs to be approved by an Manager, like a signature on a paper form?
-// Then we could have a status field in the EmployeeForm with the following states:
-//DRAFT -> SUBMITTED -> APPROVED -> COMPLETED -> REJECTED
-public class EmploymentForm {
-}
