--- src/main/java/org/example/cyberwatch/model/entitys/forms/ReportForm.java
@@ -1,113 +1,0 @@
-package org.example.cyberwatch.model.entitys.forms;
-
-import jakarta.persistence.*;
-import org.example.cyberwatch.model.enums.Department;
-import org.example.cyberwatch.model.enums.IssueType;
-import org.example.cyberwatch.model.enums.Priority;
-import org.example.cyberwatch.model.enums.Status;
-
-import java.time.LocalDateTime;
-
-@Entity
-public class ReportForm {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    Long id;
-
-    String title;
-
-    String description;
-
-    LocalDateTime reportDate;
-
-    LocalDateTime issueFound;
-
-    @Enumerated(EnumType.STRING)
-    Priority priority;
-
-    @Enumerated(EnumType.STRING)
-    Status status;
-
-    @Enumerated(EnumType.STRING)
-    IssueType issueType;
-
-    public ReportForm() {
-
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public LocalDateTime getIssueFound() {
-        return issueFound;
-    }
-
-    public void setIssueFound(LocalDateTime issueFound) {
-        this.issueFound = issueFound;
-    }
-
-    public IssueType getIssueType() {
-        return issueType;
-    }
-
-    public void setIssueType(IssueType issueType) {
-        this.issueType = issueType;
-    }
-
-    public Priority getPriority() {
-        return priority;
-    }
-
-    public void setPriority(Priority priority) {
-        this.priority = priority;
-    }
-
-    public LocalDateTime getReportDate() {
-        return reportDate;
-    }
-
-    public void setReportDate(LocalDateTime reportDate) {
-        this.reportDate = reportDate;
-    }
-
-    public Status getStatus() {
-        return status;
-    }
-
-    public void setStatus(Status status) {
-        this.status = status;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    @Enumerated(EnumType.STRING)
-    Department department;
-
-    public Department getDepartment() {
-        return department;
-    }
-
-    public void setDepartment(Department department) {
-        this.department = department;
-    }
-}
