--- src/main/java/org/example/cyberwatch/model/enums/Department.java
@@ -1,7 +1,0 @@
-package org.example.cyberwatch.model.enums;
-
-public enum Department {
-    HR,
-    MANAGEMENT,
-    CONSULTANT
-}
