--- src/main/java/org/example/cyberwatch/model/enums/IssueType.java
@@ -1,12 +1,0 @@
-package org.example.cyberwatch.model.enums;
-
-public enum IssueType {
-    HARDWARE,
-    SOFTWARE,
-    NETWORK,
-    ACCESS,
-    EMAIL,
-    PRINTER,
-    SECURITY,
-    OTHER
-}
