--- src/main/java/org/example/cyberwatch/model/enums/Priority.java
@@ -1,8 +1,0 @@
-package org.example.cyberwatch.model.enums;
-
-public enum Priority {
-    LOW,
-    MEDIUM,
-    HIGH,
-    CRITICAL,
-}
