--- src/main/java/org/example/cyberwatch/model/enums/Status.java
@@ -1,12 +1,0 @@
-package org.example.cyberwatch.model.enums;
-
-//Förslag: DRAFT -> SUBMITTED -> IN_PROGRESS -> RESOLVED -> CLOSED + REOPENED
-public enum Status {
-    DRAFT,
-    SUBMITTED,
-    IN_PROGRESS,
-    WAITING_FOR_USER, // Används när man väntar på mer information etc.
-    RESOLVED,
-    CLOSED,
-    REOPENED
-}
