--- src/main/java/org/example/cyberwatch/repository/FormRepository.java
@@ -1,4 +1,0 @@
-package org.example.cyberwatch.repository;
-
-public interface FormRepository {
-}
