--- src/main/java/org/example/cyberwatch/repository/StaffRepository.java
@@ -1,8 +1,0 @@
-package org.example.cyberwatch.repository;
-
-import org.example.cyberwatch.model.entitys.staff.Staff;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface StaffRepository extends JpaRepository<Staff, Long> {
-}
-
