--- src/main/java/org/example/cyberwatch/service/FormService.java
@@ -1,4 +1,0 @@
-package org.example.cyberwatch.service;
-
-public class FormService {
-}
