--- src/main/java/org/example/cyberwatch/service/TicketService.java
@@ -1,4 +1,0 @@
-package org.example.cyberwatch.service;
-
-public class TicketService {
-}
