+++ src/main/java/org/example/cyberwatch/shared/model/enums/Department.java
@@ -1,0 +1,7 @@
+package org.example.cyberwatch.shared.model.enums;
+
+public enum Department {
+    HR,
+    MANAGEMENT,
+    CONSULTANT
+}
