+++ src/main/java/org/example/cyberwatch/shared/model/enums/IssueType.java
@@ -1,0 +1,12 @@
+package org.example.cyberwatch.shared.model.enums;
+
+public enum IssueType {
+    HARDWARE,
+    SOFTWARE,
+    NETWORK,
+    ACCESS,
+    EMAIL,
+    PRINTER,
+    SECURITY,
+    OTHER
+}
