+++ src/main/java/org/example/cyberwatch/shared/model/enums/Priority.java
@@ -1,0 +1,8 @@
+package org.example.cyberwatch.shared.model.enums;
+
+public enum Priority {
+    LOW,
+    MEDIUM,
+    HIGH,
+    CRITICAL,
+}
